secutiry added to events views and profile views

+++ events/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Event
 
@@ -40,8 +41,13 @@ def event_detail(request, event_id):
     return render(request, 'events/event_detail.html', context)
 
 
+@login_required
 def add_event(request):
     """ Create an event """
+    if not request.user.is_superuser:
+            messages.error(request, 'Sorry, only event creators can do that.')
+            return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES)
         if form.is_valid():
@@ -61,8 +67,13 @@ def add_event(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_event(request, event_id):
     """ Edit an event """
+    if not request.user.is_superuser:
+            messages.error(request, 'Sorry, only event creators can do that.')
+            return redirect(reverse('home'))
+
     event = get_object_or_404(Event, pk=event_id)
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES, instance=event)
@@ -85,8 +96,13 @@ def edit_event(request, event_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_event(request, event_id):
     """ Delete an event """
+    if not request.user.is_superuser:
+            messages.error(request, 'Sorry, only event creators can do that.')
+            return redirect(reverse('home'))
+            
     event = get_object_or_404(Event, pk=event_id)
     event.delete()
     messages.success(request, 'Event removed!')
